Revert "refactor: shared/ui 폴더 구조 정리 및 barrel export 제거 (#232)"
This reverts commit df83e25fcfebfe26461dab4e3822beaf13f47135.
