+++ omechu-app/src/shared/ui/MainLoading.tsx
