+++ omechu-app/src/shared/ui/ProgressBar.tsx
