+++ omechu-app/src/shared/ui/SearchBar.tsx
