+++ omechu-app/src/shared/ui/auth/index.ts
@@ -1,0 +1,1 @@
+export { AuthButton } from "./AuthButton";
