+++ omechu-app/src/shared/ui/terms/index.ts
@@ -1,0 +1,1 @@
+export { default as TermsContent } from "./TermsContent";
